requirements.txt
astroid==2.12.13
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
dbus-python==1.3.2
dill==0.3.6
idna==3.4
isort==5.11.3
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.6.0
plyer==2.1.0
pycodestyle==2.10.0
pyler==0.2.0
pylint==2.15.9
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.13
wrapt==1.14.1
